Stripped back projections to bare minimum
