+++ src/nuis/measurement/Projection.h
@@ -23,19 +23,18 @@
 #include <memory>
 #include <Eigen/Dense>
 
-
 #include "yaml-cpp/yaml.h"
 using namespace YAML;
 
 #include "nuis/measurement/Variables.h"
 #include "nuis/measurement/Document.h"
 
-// NUISANCE needs some definition of errors
-// as norm need to be handled differently
-enum ErrorSource {
-     kErrorStatistical = 1,
-     kErrorSystematic,
-     kErrorNormalisation,
+enum likelihood_form {
+     kFullChi2,
+     kShapeChi2,
+     kNormChi2,
+     kShapePlusNormChi2,
+     kEvents
 };
 
 namespace nuis {
@@ -47,121 +46,38 @@ class Projection {
 
      explicit Projection(YAML::Node config);
 
-     Projection(std::string iname,
-          const Document& in_document,
-          const std::vector<Variables>& in_independent_variables,
-          const std::vector<Variables>& in_dependent_variables,
-          const std::vector<Variables>& in_dependent_covariances,
-          const std::vector<Variables>& in_independent_covariances);
-
      ~Projection();
 
-     std::string Summary();
-     void Print();
+     void Fill(int signal,
+          const std::vector<double>& indep_vals, double w = 1.0);
 
      void Scale(double v);
 
      void Reset();
 
-     int GetBin(const std::vector<double>& indep_vals);
-     int FillBin(const std::vector<double>& indep_vals, double w = 1.0);
-     int FillBin(int i, double w = 1.0);
-
-     int FillBinFromIndex(int i, double w = 1.0) {
-          return FillBin(i, w);
-     }
-
-     int FillBinFromProjection(const std::vector<double>& indep_vals,
-          double w = 1.0) {
-          return FillBin(indep_vals, w);
-     }
-
-     void ResetBin(int index);
-     void ResetBins();
-     
-     uint32_t GetMCCounts(const uint32_t i);
-     double GetMCWeight(const uint32_t i);
-     double GetMCError(const uint32_t i);
-
-     double GetTotalMCCounts();
-     double GetTotalMCWeight();
-     double GetTotalMCValue();
-
-     double GetBinContent(int index);
-     double GetBinEntries(int index);
-
-     // The bin error needs a way to select a source
-     double GetBinError(int index);
-
-     void SetTally(double v);
-
-     std::vector<double> GetSlice(
-          const std::vector<std::vector<double>>& slice,
-          const int i);
-
-
-     // Repeated functions below define a 
-     // a more accessible interface for interfacing
-     // plots with different drawing tools.
-     std::vector<double> GetXCenter();
-     std::vector<double> GetYCenter();
-     std::vector<double> GetZCenter();
-
-     std::vector<double> GetXEdge(bool low = true);
-     std::vector<double> GetYEdge(bool low = true);
-     std::vector<double> GetZEdge(bool low = true);
-
-     std::vector<double> GetXWidth();
-     std::vector<double> GetYWidth();
-     std::vector<double> GetZWidth();
-
-     std::vector<double> GetXErr();
-     std::vector<double> GetYErr();
-     std::vector<double> GetZErr();
-
-     std::vector<double> GetMC();
-     std::vector<double> GetMCErr();
-
-
-
-     double Metadata;
-
      // New bin structure scheme
-     std::string label;
-     std::string name;
      std::string title;
+     std::vector<std::string> axis_label;
+     std::vector<std::string> sub_label;
 
-     // Document document;
-     // std::vector<Variables> dependent_variables;
-     // std::vector<Variables> independent_variables;
+     std::string likelihood_type;
 
-     std::vector<std::vector<double>> bin_extent_low;
-     std::vector<std::vector<double>> bin_extent_high;
-     std::vector<std::vector<double>> bin_center;
-     std::vector<std::vector<double>> bin_width;
+     Eigen::ArrayXi bin_index;
+     Eigen::ArrayXi bin_mask;
 
-     std::vector<int>  bin_index;
-     std::vector<bool> bin_mask;
+     Eigen::ArrayXd bin_extent_low;
+     Eigen::ArrayXd bin_extent_high;
 
-     std::vector<double> data_value;
-     std::vector<double> data_error;
-     
-     std::vector<uint32_t> mc_counts;
-     std::vector<double>   mc_weights;
-     std::vector<double>   mc_errors;
+     Eigen::ArrayXd mc_count;
+     Eigen::ArrayXd mc_value;
+     Eigen::ArrayXd mc_error;
 
-     std::vector<std::vector<double>> data_covariance;
+     Eigen::ArrayXd data_value;
+     Eigen::ArrayXd data_error;
 
-
-
-     // Eigen::VectorXd 
-     // Eigen::MatrixXd bin_extent_low;
-     // Eigen::MatrixXd bin_extent_high;
-
+     Eigen::ArrayXd data_covariance;
 
 };
-
-
 
 }  // namespace measurement
 }  // namespace nuis
